+++ src/actions/article/index.ts
@@ -1,0 +1,1 @@
+export { default as refresh } from "./refresh";
