+++ src/middlewares/dev.ts
@@ -1,11 +1,11 @@
+import { ProductionModeProhibitedError } from "@/utils/auth/error";
 import { NextRequest, NextResponse } from "next/server";
-import ProductionModeProhibitedError from "@/utils/auth/errors/ProductionModeProhibitedError";
 
 // 개발 환경에서만 접근 가능하다.
 const development = process.env.NODE_ENV == "development";
 
 export async function middleware(request: NextRequest, response: NextResponse) {
-  if (!development) throw new ProductionModeProhibitedError(401);
+  if (!development) throw new ProductionModeProhibitedError(403);
   return response;
 }
