+++ src/utils/auth/error/AuthenticationError.ts
