+++ src/utils/auth/error/HTTPError.ts
