+++ src/utils/auth/error/PasswordEncryptionError.ts
