+++ src/utils/auth/error/UserAlreadyAuthenticatedError.ts
