+++ src/utils/auth/error/UserAlreadyRegisteredError.ts
