+++ src/utils/auth/error/UserNotAuthenticatedError.ts
