+++ src/utils/auth/error/index.ts
@@ -1,0 +1,14 @@
+export { default as AuthenticationError } from "./AuthenticationError";
+export { default as InvalidSecretError } from "./InvalidSecretError";
+export { default as InvalidTokenError } from "./InvalidTokenError";
+export { default as InvalidUserError } from "./InvalidUserError";
+export { default as PasswordEncryptionError } from "./PasswordEncryptionError";
+export { default as ProductionModeProhibitedError } from "./ProductionModeProhibitedError";
+export { default as SaltCreationError } from "./SaltCreationError";
+export { default as TokenCreationError } from "./TokenCreationError";
+export { default as TokenValidationError } from "./TokenValidationError";
+export { default as UserAlreadyAuthenticatedError } from "./UserAlreadyAuthenticatedError";
+export { default as UserAlreadyRegisteredError } from "./UserAlreadyRegisteredError";
+export { default as UserNotAuthenticatedError } from "./UserNotAuthenticatedError";
+export { default as UserNotFoundError } from "./UserNotFoundError";
+export { default as HTTPError } from "./HTTPError";
